CDKBootstrap-Advanced/src/requirements.txt
boto3
aws-cdk.core
